requirements.txt
qdrant-client
openai
SPARQLWrapper
tqdm
python-dotenv